requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
faiss-cpu==1.7.4
google-generativeai==0.3.2
grpcio==1.60.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
numpy==1.26.2
requests==2.32.5
pickle5==0.0.12

